requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.0
chromadb==0.4.18
tiktoken==0.5.1
lxml==4.9.3
aiohttp==3.9.1

